requirements.txt
#requirements.txt
python-telegram-bot[job-queue]>=20.8
aiohttp==3.9.3
python-dotenv==1.0.1
APScheduler==3.10.4
